init response

+++ src/main/java/com/tcdt/qlnvkhoachphi/response/chitieukehoachnam/kehoachluongthucdutru/tondaunam/TonDauNamRes.java
@@ -5,13 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class TonDauNamRes {
-	private BigDecimal tongSoQuyThoc;
+	private Double tongSoQuyThoc;
 	private TrongDoTonDauNamRes trongDo;
 }
